requirements.txt
python-dotenv
requests
pandas
plotly
dash
pick